Seperate effects_v1 file
